+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [0.8.16] - 2026-05-15
+
+### Added
+
+- feat(api): prevent loopback device approval URLs for remote APIs (abbb6296e27f)
+
+### Dependencies
+
+- Release @treeseed/agent 0.8.16.
+
 ## [0.8.15] - 2026-05-15
 
 ### Added